(rename) radish plugin module
