+++ core/exports/plugins.ts
@@ -2,5 +2,5 @@ export { pluginStripTypes } from "../src/plugins/strip-types.ts";
 export { pluginConfig } from "../src/plugins/config.ts";
 export { pluginImportmap } from "../src/plugins/importmap.ts";
 export { pluginIO } from "../src/plugins/io.ts";
-export { pluginRadish } from "../src/plugins/plugins.ts";
+export { pluginRadish } from "../src/plugins/radish.ts";
 export { pluginManifest, updateManifest } from "../src/plugins/manifest.ts";
